Thot Police Version 2
Version 2 is completely arbitrary, it simply denotes that I gave up on my first attempt (whose code is still included in the repository) and started everything over from scratch in the "v2" folder. I currently am trying to figure out how to make the program read from a text file to find the whitelisted and blacklisted words instead of leaving them hard coded in the middle of the program. All my testing to try to figure this out is located in the "node testing stuff" because I didn't quite know what else to call it, but the two js files in there, v1 and v2 were my two attempts at sorting the text in the files. nodeTestingv2 is the current solution, but I can't quite figure out how to make the function return the text array, I think it has to do with the fact that the file read is asynchonous and I have no idea what that means.